retail-lakehouse-project/requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
tensorflow
plotly
ipykernel
jupyter
